requirements.txt
kafka-python
click
